Create Union_Find.h
This header file contains the following two structure definitions
a. A Node struct : This can be used to hold an individual element belonging to any Disjoint Subset
b. A disjointSetList struct : This is the container that holds all the Disjoint Subsets created till date

It also contains a class definition named disjointSet that declares all the methods available to modify the Union-Find Data Structure. These methods are
a. makeSet : Used to create a Singleton set
b. findInSet : Find the Disjoint Subset that contains a particular element
c. unionOfSets : Perform the union of two disjoint subsets
d. insertIntoContainer : A helper function used to insert a Singleton set in to a container